pyproject.toml
[tool.poetry]
name = "brainly-moderation-dashboard-api"
version = "0.1.0"
description = "API of the ultimate moderation dashboard for Brainly moderators"
authors = ["vlaex <https://t.me/vlaex>"]
license = "MIT"
readme = "README.md"
packages = [{include = "brainly_moderation_dashboard_api"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.2"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.6"
django-environ = "^0.10.0"
django-timescaledb = "^0.2.13"
django-cors-headers = "^4.1.0"
djangorestframework-simplejwt = "^5.2.2"
httpx = {extras = ["http2"], version = "^0.24.1"}
python-graphql-client = "^0.4.3"
gunicorn = "^20.1.0"
django-debug-toolbar = "^4.1.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
